working on GPU kernels

+++ include/CppNet/regularizations/regularizations.hpp
@@ -18,59 +18,59 @@ namespace CppNet
             virtual ~Regularizer() = default;
 
             // Apply regularization update to parameters (stub)
-            virtual void apply(Eigen::Tensor<double, 2>& params) = 0;
+            virtual void apply(Eigen::Tensor<float, 2>& params) = 0;
 
             // Return penalty term for loss
-            virtual double penalty(const Eigen::Tensor<double, 2>& params) = 0;
+            virtual float penalty(const Eigen::Tensor<float, 2>& params) = 0;
         };
 
         /************************************** L1 Regularization *************************************/
         class L1 : public Regularizer
         {
         private:
-            double lambda;
+            float lambda;
 
         public:
-            L1(double lambda = 0.01);
-            void apply(Eigen::Tensor<double, 2>& params) override;
-            double penalty(const Eigen::Tensor<double, 2>& params) override;
+            L1(float lambda = 0.01f);
+            void apply(Eigen::Tensor<float, 2>& params) override;
+            float penalty(const Eigen::Tensor<float, 2>& params) override;
         };
 
         /************************************** L2 Regularization *************************************/
         class L2 : public Regularizer
         {
         private:
-            double lambda;
+            float lambda;
 
         public:
-            L2(double lambda = 0.01);
-            void apply(Eigen::Tensor<double, 2>& params) override;
-            double penalty(const Eigen::Tensor<double, 2>& params) override;
+            L2(float lambda = 0.01);
+            void apply(Eigen::Tensor<float, 2>& params) override;
+            float penalty(const Eigen::Tensor<float, 2>& params) override;
         };
 
         /************************************** ElasticNet (L1 + L2) *************************************/
         class ElasticNet : public Regularizer
         {
         private:
-            double l1;
-            double l2;
+            float l1;
+            float l2;
 
         public:
-            ElasticNet(double l1 = 0.01, double l2 = 0.01);
-            void apply(Eigen::Tensor<double, 2>& params) override;
-            double penalty(const Eigen::Tensor<double, 2>& params) override;
+            ElasticNet(float l1 = 0.01, float l2 = 0.01);
+            void apply(Eigen::Tensor<float, 2>& params) override;
+            float penalty(const Eigen::Tensor<float, 2>& params) override;
         };
 
         /************************************** Dropout *************************************/
         class Dropout : public Regularizer
         {
         private:
-            double rate;
+            float rate;
 
         public:
-            Dropout(double rate = 0.5);
-            void apply(Eigen::Tensor<double, 2>& params) override;
-            double penalty(const Eigen::Tensor<double, 2>& params) override;
+            Dropout(float rate = 0.5);
+            void apply(Eigen::Tensor<float, 2>& params) override;
+            float penalty(const Eigen::Tensor<float, 2>& params) override;
         };
     }
 }
